+++ NEWS.md
@@ -8,6 +8,7 @@
 
 ### folio-spring-kafka
 * [FOLSPRINGS-232](https://folio-org.atlassian.net/browse/FOLSPRINGS-232) Add tenant-aware Kafka message filtering support for Spring modules
+* [FOLSPRINGS-233](https://folio-org.atlassian.net/browse/FOLSPRINGS-233) Cache Kafka tenant entitlement lookups instead of calling the sidecar per message
 
 ## 10.0.0 2026-04-06